+++ src/Pages/SettingsPageModel.cs
@@ -25,7 +25,7 @@ public partial class SettingsPageModel : ObservableObject
     readonly WeakReference<SettingsPage> _weakPage;
     readonly DLSSSettingsManager _dlssSettingsManager;
     public string CurrentLogPath => Logger.GetCurrentLogPath();
-    public string AppVersion => $"{App.CurrentApp.GetVersionString()} / 1.0.2+";
+    public string AppVersion => $"{App.CurrentApp.GetVersionString()} / 1.0.3+";
 
     [ObservableProperty]
     public partial ComboBoxOption SelectedDlssOnScreenIndicator { get; set; }